It's not a dataframe so don't use df as the variable name
It was a temporary value so doesn't really matter, but df is almost
the defacto name for a dataframe so avoid needless confusion.

Currently ignored isomer lines in the NUBASE file return an empty
dictionary, but this was being converted to a row of NaNs in the
dataframe and messing up the type. Filter them out before converting
everything into a dataframe.